pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "refactorio"
version = "0.1.0"
description = "Automated whole-repo refactoring orchestrator using Claude Code CLI"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Refactorio Team" }
]
keywords = ["refactoring", "automation", "claude", "code"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.5.0",
    "gitpython>=3.1.0",
    "jsonschema>=4.20.0",
]

[project.optional-dependencies]
gui = [
    "customtkinter>=5.2.0",
    "ttkbootstrap>=1.10.0",
    "pygments>=2.17.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
refactorio = "refactor_bot.main:app"

[project.urls]
Homepage = "https://github.com/refactorio/refactorio"
Documentation = "https://github.com/refactorio/refactorio#readme"
Repository = "https://github.com/refactorio/refactorio"

[tool.hatch.build.targets.wheel]
packages = ["refactor_bot"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
